Rename wrong-quotes-for-template-string to no-template-curly-in-string
