+++ lib/rules/no-template-curly-in-string.js
